Add FirstOrNone extension method for IEnumerable<T>
